Switching between Project List and Posts working
