+++ frontend/src/Components/CommunityProjects/Projects.jsx
@@ -4,28 +4,11 @@ import './Projects.css';
 
 // eslint-disable-next-line react/prefer-stateless-function
 class Projects extends React.Component {
-  // constructor(props) {
-  //   super(props);
-  //   // this.onClickPosts = this.onClickPosts.bind(this);
-  // }
-
-  // onClickPosts() {
-  //   // const index = parseInt(this.index, 10);
-  //   this.setState({
-  //     displayPosts: true
-  //   });
-  // }
-
   render() {
-    const { title, desc } = this.props;
+    const { title, desc, onClickPosts } = this.props;
     return (
       <div className="project">
-        <div
-          className="projectText"
-          role="presentation"
-          // onClick={onClickPosts}
-          // onClick={this.setRedirect}
-        >
+        <div className="projectText" role="presentation" onClick={onClickPosts}>
           <div className="projectTitle">
             <span>{title}</span>
           </div>
@@ -42,7 +25,7 @@ class Projects extends React.Component {
 Projects.propTypes = {
   title: PropTypes.string.isRequired,
   desc: PropTypes.string.isRequired,
-  // onClickPosts: PropTypes.func.isRequired,
+  onClickPosts: PropTypes.func.isRequired,
 };
 
 export default Projects;
