--- frontend/src/Components/NewPost/NewPost.css
@@ -1,8 +1,0 @@
-.newPost {
-  height: 50px;
-  width: 90vw;
-}
-
-.button {
-  background-color: #FF784E;
-}
